# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Central Command Console API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CommandInput {
  /** Raw command string typed by the operator */
  command: string;
}

export type CommandResultType = typeof CommandResultType[keyof typeof CommandResultType];


export const CommandResultType = {
  info: 'info',
  success: 'success',
  error: 'error',
  system: 'system',
  clear: 'clear',
} as const;

export interface CommandResult {
  ok: boolean;
  command?: string;
  output: string;
  type: CommandResultType;
  timestamp?: string;
}

export type ModuleStatusStatus = typeof ModuleStatusStatus[keyof typeof ModuleStatusStatus];


export const ModuleStatusStatus = {
  online: 'online',
  offline: 'offline',
  degraded: 'degraded',
} as const;

export interface ModuleStatus {
  name: string;
  status: ModuleStatusStatus;
  /** @nullable */
  latency: number | null;
}

export interface SystemStatus {
  /** CPU usage percentage */
  cpu: number;
  /** Memory usage percentage */
  memory: number;
  /** Server uptime in seconds */
  uptime: number;
  modules: ModuleStatus[];
  timestamp: string;
}

export interface PhoneInput {
  phone: string;
}

export type PhoneResultRisk = typeof PhoneResultRisk[keyof typeof PhoneResultRisk];


export const PhoneResultRisk = {
  safe: 'safe',
  warning: 'warning',
  danger: 'danger',
} as const;

export interface PhoneResult {
  phone: string;
  risk: PhoneResultRisk;
  label: string;
  details: string;
  /** Number of reported incidents */
  calls: number;
  /** @nullable */
  lastSeen?: string | null;
}

export interface EmailInput {
  email: string;
}

export type BreachItemSeverity = typeof BreachItemSeverity[keyof typeof BreachItemSeverity];


export const BreachItemSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface BreachItem {
  name: string;
  date: string;
  dataTypes: string[];
  severity: BreachItemSeverity;
}

export interface EmailResult {
  email: string;
  compromised: boolean;
  /** 0-100 risk score */
  riskScore: number;
  breaches: BreachItem[];
  /** @nullable */
  recommendation?: string | null;
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  INFO: 'INFO',
  WARN: 'WARN',
  ALERT: 'ALERT',
  OK: 'OK',
} as const;

export interface LogEntry {
  id: string;
  level: LogEntryLevel;
  message: string;
  timestamp: string;
}