# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Central Command Console API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: command
    description: Command terminal operations
  - name: cybersentry
    description: CyberSentry anti-fraud and security operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /command:
    post:
      operationId: executeCommand
      tags: [command]
      summary: Execute a terminal command
      description: Dispatches a command string and returns the formatted response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandInput"
      responses:
        "200":
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          description: Unknown command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /system/status:
    get:
      operationId: getSystemStatus
      tags: [command]
      summary: Get live system status
      description: Returns current CPU, memory, uptime, and module statuses
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"
  /cybersentry/phone:
    post:
      operationId: checkPhone
      tags: [cybersentry]
      summary: Check phone number reputation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneInput"
      responses:
        "200":
          description: Phone reputation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneResult"
  /cybersentry/email:
    post:
      operationId: scanEmail
      tags: [cybersentry]
      summary: Scan email for data breaches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Email breach scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailResult"
  /cybersentry/log:
    get:
      operationId: getSecurityLog
      tags: [cybersentry]
      summary: Get live security event log
      responses:
        "200":
          description: List of security log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CommandInput:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          description: Raw command string typed by the operator
    CommandResult:
      type: object
      required:
        - ok
        - output
        - type
      properties:
        ok:
          type: boolean
        command:
          type: string
        output:
          type: string
        type:
          type: string
          enum: [info, success, error, system, clear]
        timestamp:
          type: string
    SystemStatus:
      type: object
      required:
        - cpu
        - memory
        - uptime
        - modules
        - timestamp
      properties:
        cpu:
          type: number
          description: CPU usage percentage
        memory:
          type: number
          description: Memory usage percentage
        uptime:
          type: integer
          description: Server uptime in seconds
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleStatus"
        timestamp:
          type: string
    ModuleStatus:
      type: object
      required:
        - name
        - status
        - latency
      properties:
        name:
          type: string
        status:
          type: string
          enum: [online, offline, degraded]
        latency:
          type: ["integer", "null"]
    PhoneInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    PhoneResult:
      type: object
      required: [phone, risk, label, details, calls]
      properties:
        phone:
          type: string
        risk:
          type: string
          enum: [safe, warning, danger]
        label:
          type: string
        details:
          type: string
        calls:
          type: integer
          description: Number of reported incidents
        lastSeen:
          type: ["string", "null"]
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    BreachItem:
      type: object
      required: [name, date, dataTypes, severity]
      properties:
        name:
          type: string
        date:
          type: string
        dataTypes:
          type: array
          items:
            type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
    EmailResult:
      type: object
      required: [email, compromised, riskScore, breaches]
      properties:
        email:
          type: string
        compromised:
          type: boolean
        riskScore:
          type: integer
          description: 0-100 risk score
        breaches:
          type: array
          items:
            $ref: "#/components/schemas/BreachItem"
        recommendation:
          type: ["string", "null"]
    LogEntry:
      type: object
      required: [id, level, message, timestamp]
      properties:
        id:
          type: string
        level:
          type: string
          enum: [INFO, WARN, ALERT, OK]
        message:
          type: string
        timestamp:
          type: string