# lib/api-zod/src/generated/types/breachItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Central Command Console API
 * OpenAPI spec version: 0.1.0
 */
import type { BreachItemSeverity } from './breachItemSeverity';

export interface BreachItem {
  name: string;
  date: string;
  dataTypes: string[];
  severity: BreachItemSeverity;
}