# lib/api-zod/src/generated/types/emailResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Central Command Console API
 * OpenAPI spec version: 0.1.0
 */
import type { BreachItem } from './breachItem';

export interface EmailResult {
  email: string;
  compromised: boolean;
  /** 0-100 risk score */
  riskScore: number;
  breaches: BreachItem[];
  /** @nullable */
  recommendation?: string | null;
}