# lib/api-zod/src/generated/types/logEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Central Command Console API
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntryLevel } from './logEntryLevel';

export interface LogEntry {
  id: string;
  level: LogEntryLevel;
  message: string;
  timestamp: string;
}