# lib/api-zod/src/generated/types/phoneResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Central Command Console API
 * OpenAPI spec version: 0.1.0
 */
import type { PhoneResultRisk } from './phoneResultRisk';

export interface PhoneResult {
  phone: string;
  risk: PhoneResultRisk;
  label: string;
  details: string;
  /** Number of reported incidents */
  calls: number;
  /** @nullable */
  lastSeen?: string | null;
}